Start lifting allow_missing for rest in vector server
